# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control System API",
        "description": "API documentation for Risk Control System",
        "contact": {
            "email": "admin@riskcontrol.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/trades": {
            "get": {
                "tags": [
                    "Trades"
                ],
                "summary": "Get list of trades",
                "description": "Returns list of trades with optional filtering",
                "operationId": "getTradesList",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "query",
                        "description": "Filter by Account ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by Status (OPEN/CLOSED)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Trades"
                ],
                "summary": "Store new trade",
                "description": "Returns trade data",
                "operationId": "storeTrade",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTradeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/trades/{id}": {
            "get": {
                "tags": [
                    "Trades"
                ],
                "summary": "Get trade information",
                "description": "Returns trade data",
                "operationId": "getTradeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Trade ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Trades"
                ],
                "summary": "Update existing trade",
                "description": "Returns updated trade data",
                "operationId": "updateTrade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Trade ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTradeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreTradeRequest": {
                "title": "Store Trade Request",
                "description": "Store Trade request body data",
                "required": [
                    "account_id",
                    "type",
                    "volume",
                    "open_time",
                    "open_price",
                    "status"
                ],
                "properties": {
                    "account_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "BUY",
                            "SELL"
                        ],
                        "example": "BUY"
                    },
                    "volume": {
                        "type": "number",
                        "format": "float",
                        "example": 1.5
                    },
                    "open_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T10:00:00Z"
                    },
                    "close_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T12:00:00Z",
                        "nullable": true
                    },
                    "open_price": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "close_price": {
                        "type": "number",
                        "format": "float",
                        "example": 155,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "CLOSED"
                        ],
                        "example": "OPEN"
                    }
                },
                "type": "object"
            },
...
            "UpdateTradeRequest": {
                "title": "Update Trade Request",
                "description": "Update Trade request body data",
                "properties": {
                    "close_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T12:00:00Z"
                    },
                    "close_price": {
                        "type": "number",
                        "format": "float",
                        "example": 155
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "CLOSED"
                        ],
                        "example": "CLOSED"
                    }
                },
                "type": "object"
            }
        }
    },
...
}